requirements.txt
# To regenerate this list, use pip install pipreqs && pipreqs /path/to/bmi3d --print

Cython==0.29.22
Django==4.1
geometry==0.0.23
h5py==3.10.0
matplotlib==3.8.3
namelist==0.1.0
nitime==0.9
numpy==1.26.4
Owl==0.2.1
paho_mqtt==1.5.1
pandas==1.5.3
Phidgets==2.1.8
PyAutoGUI==0.9.50
pyFirmata==1.1.0
pygame==2.0.0
PyJWT==2.8.0
PyLink==0.3.3
PyMySQL==1.0.2
PyOpenGL==3.1.7
pyserial==3.5
pyzmq==22.1.0
QtPy==1.9.0
Requests==2.31.0
scikit_learn==0.24.2
scipy==1.5.4
setuptools==65.5.0
statsmodels==0.14.1
tables==3.9.2
tabulate==0.9.0
tasks==2.5.0
tornado==6.1
traits==6.1.0
urllib3==1.26.13
msgpack==0.5.6
pyopenxr==1.1
